server/requirements.txt
uvicorn
fastapi
beanie
motor
redis
rq
python-dotenv
pymupdf
docling
pinecone
neo4j
langchain
langchain_neo4j
langchain_huggingface
langchain-google-genai
langchain-mistralai
langchain_experimental
google-generativeai
mistralai
sentence-transformers